image-classification-tiny-image-net/requirements.txt
lightning==2.1.0
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
scikit_learn==1.3.2
timm==0.9.8
torch==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
